backend/python-ml-api/requirements.txt
fastapi
uvicorn[standard]
pydantic
mysql-connector-python
requests
scikit-learn
numpy
pandas
python-dotenv
google-generativeai
xgboost
